refactor: di 모듈에서 안드로이드 의존성 제거
