+++ app/build.gradle.kts
@@ -73,7 +73,7 @@ dependencies {
     testImplementation("io.mockk:mockk:1.13.5")
     androidTestImplementation("io.mockk:mockk-android:1.13.5")
     // DI Library
-    implementation(project(":di"))
+    implementation(project(":bunadi"))
 }
 
 kapt {
