+++ app/src/main/java/woowacourse/shopping/ui/common/di/module/RepositoryModule.kt
@@ -2,11 +2,10 @@ package woowacourse.shopping.ui.common.di.module
 
 import android.content.Context
 import androidx.room.Room
-import com.buna.di.module.Module
 import woowacourse.shopping.data.CartProductDao
 import woowacourse.shopping.data.ShoppingDatabase
 
-class RepositoryModule(private val context: Context) : Module {
+class RepositoryModule(private val context: Context) : com.woowacourse.bunadi.module.Module {
     fun provideCartProductDao(): CartProductDao = Room.databaseBuilder(
         context,
         ShoppingDatabase::class.java,
