+++ bunadi/src/main/java/com/woowacourse/bunadi/annotation/Inject.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.bunadi.annotation
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
+annotation class Inject
