+++ bunadi/src/main/java/com/woowacourse/bunadi/module/Module.kt
@@ -1,0 +1,3 @@
+package com.woowacourse.bunadi.module
+
+interface Module
