--- di/src/main/java/com/buna/di/annotation/Inject.kt
@@ -1,5 +1,0 @@
-package com.buna.di.annotation
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
-annotation class Inject
