--- di/src/main/java/com/buna/di/module/Module.kt
@@ -1,3 +1,0 @@
-package com.buna.di.module
-
-interface Module
