--- di/src/main/java/com/buna/di/util/diAssistant/SubTypeConverter.kt
@@ -1,21 +1,0 @@
-package com.buna.di.util.diAssistant
-
-import com.buna.di.injector.DependencyKey
-import kotlin.reflect.KType
-
-data class SubTypeConverter(
-    private val converter: MutableMap<DependencyKey, KType> = mutableMapOf(),
-) {
-    fun saveType(parentTypeKey: DependencyKey, childType: KType) {
-        converter[parentTypeKey] = childType
-    }
-
-    fun convertType(parentTypeKey: DependencyKey, defaultType: KType): KType {
-        return converter[parentTypeKey] ?: defaultType
-    }
-
-    fun clear(): SubTypeConverter {
-        converter.clear()
-        return copy(converter = mutableMapOf())
-    }
-}
