--- di/src/main/java/com/buna/di/util/viewModel/ViewModelFactory.kt
@@ -1,13 +1,0 @@
-package com.buna.di.util.viewModel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewmodel.initializer
-import androidx.lifecycle.viewmodel.viewModelFactory
-
-inline fun <reified VM : ViewModel> viewModelFactory(crossinline create: () -> VM): ViewModelProvider.Factory =
-    viewModelFactory {
-        initializer {
-            create()
-        }
-    }
