--- di/src/test/java/com/buna/di/injector/fakeClasses/FakeConstructorDependency.kt
@@ -1,21 +1,0 @@
-package com.buna.di.injector.fakeClasses
-
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModel
-import com.buna.di.util.viewModel.viewModel
-
-annotation class AConstructorDependencyQualifier
-
-interface ConstructorDependency
-
-@AConstructorDependencyQualifier
-class AConstructorDependency : ConstructorDependency
-
-class ConstructorTestActivity : AppCompatActivity() {
-    val viewModel: ConstructorTestViewModel by viewModel()
-}
-
-class ConstructorTestViewModel(
-    @AConstructorDependencyQualifier
-    val constructorDependency: ConstructorDependency,
-) : ViewModel()
